backend/requirements.txt
djangorestframework==3.12.4
django==3.2.7
# Backend Dependencies
python-dotenv==0.19.0
requests==2.26.0
pytest==6.2.5
gunicorn==20.1.0
psycopg2-binary==2.9.9
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow